The big 4.4BSD Lite to FreeBSD 2.0.0 (Development) patch.
Reviewed by:	Rodney W. Grimes
Submitted by:	John Dyson and David Greenman


Former-commit-id: 152b5c75f316212674f6a4fc88996e0029cc7bbc

+++ sys/netinet/tcp_timer.h
@@ -255,7 +255,7 @@ name|tvmin
 parameter_list|,
 name|tvmax
 parameter_list|)
-value|{ \ 	(tv) = (value); \ 	if ((tv)< (tvmin)) \ 		(tv) = (tvmin); \ 	else if ((tv)> (tvmax)) \ 		(tv) = (tvmax); \ }
+value|{ \ 	(tv) = (value); \ 	if ((u_long)(tv)< (u_long)(tvmin)) \ 		(tv) = (tvmin); \ 	else if ((u_long)(tv)> (u_long)(tvmax)) \ 		(tv) = (tvmax); \ }
 end_define
 
 begin_ifdef
